Make card list clear and store selected cards
The card list will now add selected cards to an array
so that actions can eventually be performed on the
database with them. In addition, switching to a mode
that views cards instead of selecting them will reset
the cards that are selected on the screen and clear them
from the selected cards array.