Card list can delete cards from decks
A button for deleting cards from decks has
been added to the card list that behaves similarly
to the add button. When a deck and cards are selected,
this button will delete the selected cards from the
deck (if they exist). Code artifacts have been removed from
the card list service.