Add e2e tests for results pop up
e2e now tests that the pop up pops up after
the got-it button has been clicked for all
the cards, and that it isn't tied to the last
card. It also checks that the play again button
goes to the deck-list page, and the home button
goes to the home page.

Review/QA #50
Review/QA #51
Review/QA #52