Create the start page
There is now a page where you can select
the deck you want to play and it's connected
to the home page play button. The deck list
page doesn't have a play button anymore and
there is now a tab on the navbar which will
go to this new start page.

#54 #55