Supplement card list e2e testing
Card list tests for adding and deleting cards
have been added. One checks that the number of cards
in a deck will increase when using the add button, another
checks that duplicate cards cannot be added to a deck, and
the final checks that cards can be deleted from the deck.
Many methods were added to the associated po files to help
make these tests work, and small changes to the html have
been made with related components, mainly the addition
of dynamic id-setting inside of lists.