Decks have editable names
The deck page now has an edit button next to
the name of the deck. When pressed, a field
with the old deck name as a placeholder
appears in place of the deck title. Two other
buttons, save changes and cancel, also appear
and the edit button disappears until the editing
phase is completed. The deck name is updated through
the service communicating with the backend. Currently,
a page refresh is required to see that the name has
successfully updated.