Add button, Dialog for deleting decks
A button for deleting decks has been added
to the deck page which, when pressed, will
open a confirmation dialog with the name of
the deck included and two buttons. One
button will cancel the deletion, which closes
the dialog, and the other will eventually delete
the deck but for now just routes to the deck-list
page. Testing has been fixed for the new dialog.