Add e2e tests for deleting decks
Two e2e tests which check the functionality
of the cancel button and delete button on
the deck page for deleting decks have been
added and are passing. The cancel test
checks that the page remains the same after
pressing cancel and the delete test checks
that the deck has been removed from the
deck list page.