Add ability to delete cards
Cards can now be deleted from the system
from the card list menu. A trash button
has been added to the pop up dialog that
appears when a card is clicked on in the card
list. When pressed, another dialog will pop
up with confirmation dialog and buttons. At
present, a page refresh is needed to see that
the changes have actually been made.