Swap homepage testing type
The unit tests for homepage were
unable to be adapted with the auth
service for some unknown reason, so all
five would either fail or not run. To
address this, these tests have been turned
into e2e tests and the unit tests for the
home component have been deprecated. We think
this is justified because the same functionality
is tested in the new e2e tests, and because
no methods or instance variables were actually
being tested in the old unit tests.